refactor: zmiana typu danych porównywalnej wartości liczbowej statystyki na float

+++ Squadra.Server/Modules/Statystyki/DTO/StatystykaDTO.cs
@@ -4,7 +4,7 @@ public record StatystykaDTO(
     int Id,
     string Nazwa,
     string Wartosc,
-    int? PorownywalnaWartoscLiczbowa,
+    double? PorownywalnaWartoscLiczbowa,
     int KategoriaId,
     string KategoriaNazwa,
     int? RolaId,
